fix interface code.

+++ frameworks/native/sms/src/send_short_message_callback_stub.cpp
@@ -33,7 +33,7 @@ int SendShortMessageCallbackStub::OnRemoteRequest(
         return TELEPHONY_ERR_DESCRIPTOR_MISMATCH;
     }
     switch (code) {
-        case SendShortMessageCallbackInterfaceCode::ON_SMS_SEND_RESULT: {
+        case static_cast<int>(SendShortMessageCallbackInterfaceCode::ON_SMS_SEND_RESULT): {
             int32_t result = data.ReadInt32();
             OnSmsSendResult(static_cast<ISendShortMessageCallback::SmsSendResult>(result));
             return SMS_DEFAULT_RESULT;
